TypicalPersons: Move Index constants to TypicalIndexes
TypicalPersons contain Index constants.

This violates Single Responsibility Principle as TypicalPersons should
only contain constants and helper methods pertaining to Person objects.

Let’s add TypicalIndex and move the Index constants to this new class.
