+++ src/test/java/seedu/address/testutil/TypicalIndexes.java
@@ -1,0 +1,12 @@
+package seedu.address.testutil;
+
+import seedu.address.commons.core.index.Index;
+
+/**
+ * A utility class containing a list of {@code Index} objects to be used in tests.
+ */
+public class TypicalIndexes {
+    public static final Index INDEX_FIRST_PERSON = Index.fromOneBased(1);
+    public static final Index INDEX_SECOND_PERSON = Index.fromOneBased(2);
+    public static final Index INDEX_THIRD_PERSON = Index.fromOneBased(3);
+}
