+++ src/test/java/systemtests/DeleteCommandSystemTest.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.commands.DeleteCommand.MESSAGE_DELETE_PERSON_S
 import static seedu.address.testutil.TestUtil.getLastIndex;
 import static seedu.address.testutil.TestUtil.getMidIndex;
 import static seedu.address.testutil.TestUtil.getPerson;
-import static seedu.address.testutil.TypicalPersons.INDEX_FIRST_PERSON;
+import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import static systemtests.AppStateAsserts.assertCommandFailure;
 import static systemtests.AppStateAsserts.assertCommandSuccess;
